+++ frigate/detectors/util.py
